requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.7
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
kombu==5.5.0
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index==0.12.24
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.24.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-pinecone==0.4.5
llama-parse==0.6.4.post1
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==1.66.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
prompt_toolkit==3.0.50
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
